third_party/WebKit/LayoutTests/media/W3C/video/canPlayType/canPlayType_supported_but_no_codecs_parameter_3-expected.txt

video.canPlayType() - supported format w/o codecs parameter

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "maybe" is "maybe"
PASS successfullyParsed is true

TEST COMPLETE
spec reference
